Separate currency service.
