+++ README.md
@@ -1,0 +1,7 @@
+## Why
+Get familiar with Exonum blockchain principles. Split logic into
+meaningfull layers and modules for better future evolving.
+
+## Purpose
+Make more services (currency, vote, ...) for variety blockchain use (like smart contracts
+in some other blockchain platforms).
