+++ src/currency/mod.rs
@@ -1,3 +1,4 @@
 pub mod errors;
 pub mod schema;
 pub mod transactions;
+pub mod service;
